Update int_reg
you were missing the x parameter in the parameter description. Changed importfrom to import because otherwise it would say that it's missing. Added color as a parameter so users can change the color of the regression line. Took the imports off of the second function because i believe if its above you don't have to respecify them